fix: properly include comments after indented blocks
OverPy previously had a bug with the check for skipping comments
when finding the child lines of an indented block, which meant
that the check never skipped commented lines.

In particular, the request for `.tokens[0][0]` always returned
`undefined`, which was then typecast to true by the negation
operator. Since `true` is never strictly equal to "#", the
condition always returned true, and thus the check for indentation
levels was also always performed.

See commit 978131f for the first introduction of this behavior.